Updated dicom2stl and nii2stl files

+++ files/dicom2stl.sh
@@ -6,13 +6,18 @@ SCANNAME=${BASENAME%%.*}
 PRINTDIR=/3dprintscript
 NEWDIR=$PRINTDIR/scans/$SCANNAME
 
-mkdir -p $NEWDIR
-mv $FILENAME $NEWDIR
-unzip $NEWDIR/$BASENAME -d $NEWDIR
-rm $NEWDIR/$BASENAME
+if [[ $BASENAME == *.zip ]]
+then
+    mkdir -p $NEWDIR
+    mv $FILENAME $NEWDIR
+    unzip $NEWDIR/$BASENAME -d $NEWDIR
+    rm $NEWDIR/$BASENAME
 
-$PRINTDIR/dicom2nii.sh $SCANNAME
+    $PRINTDIR/dicom2nii.sh $SCANNAME
 
-$PRINTDIR/mri2stl.sh $SCANNAME
+    $PRINTDIR/mri2stl.sh $SCANNAME
 
-#mv $NEWDIR/output/final_s.stl $PRINTDIR/brain_$SCANNAME.stl
+    #mv $NEWDIR/output/final_s.stl $PRINTDIR/brain_$SCANNAME.stl
+else
+    echo "File is not a *.zip file."
+fi
